--- Assets/Prefabs/Animal/Goat.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 32b10f0c79bbc1342adc27ec7877a416
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
