--- Assets/Prefabs/Animal/Pig.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1c5818404b8f81949b5b768db257ca65
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
