+++ Assets/Prefabs/Grid/Grid.prefab
@@ -9,7 +9,7 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 6153418739021730139}
-  - component: {fileID: 1304596584}
+  - component: {fileID: 6875162307949251115}
   m_Layer: 0
   m_Name: Grid
   m_TagString: Untagged
@@ -31,7 +31,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &1304596584
+--- !u!114 &6875162307949251115
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -40,7 +40,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 6153418739021730138}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 63884f73dc95d0644b8a28595c2c3a40, type: 3}
+  m_Script: {fileID: 11500000, guid: c07f4ee3d3b43fc46bdde363c9ff8da5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   tile: {fileID: 291942369377142905, guid: 0c5028c134ed30c4290943b4798ee1dd, type: 3}
