+++ Assets/Prefabs/Managers/GameManager.prefab
@@ -9,9 +9,11 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 376632600952604116}
-  - component: {fileID: 376632600952604117}
-  - component: {fileID: 376632600952604118}
-  - component: {fileID: 376632600956600434}
+  - component: {fileID: 1876302422284711517}
+  - component: {fileID: 2771721594638925932}
+  - component: {fileID: 4458347510344920136}
+  - component: {fileID: 7504042938434839235}
+  - component: {fileID: 1998927541}
   m_Layer: 0
   m_Name: GameManager
   m_TagString: GameManager
@@ -33,7 +35,7 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &376632600952604117
+--- !u!114 &1876302422284711517
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -42,10 +44,10 @@ MonoBehaviour:
   m_GameObject: {fileID: 376632600952604115}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 18b3752c380f2fb4889971515faf916a, type: 3}
+  m_Script: {fileID: 11500000, guid: 30bb6efb377e23849ada76f03f6d2adc, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &376632600952604118
+--- !u!114 &2771721594638925932
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -54,12 +56,37 @@ MonoBehaviour:
   m_GameObject: {fileID: 376632600952604115}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dec7771bf9d127b49be23087f681e362, type: 3}
+  m_Script: {fileID: 11500000, guid: f77ad97dc2a7d9741a95e7d66aad4fab, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  onAttackMode: 0
+--- !u!114 &4458347510344920136
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 376632600952604115}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 1cc03ae6223af89439634f74da42f44f, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &7504042938434839235
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 376632600952604115}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2ff8cff738eb82245a5894bc9ca530e5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   animalList: []
   vegetableList: []
---- !u!114 &376632600956600434
+--- !u!114 &1998927541
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -68,6 +95,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 376632600952604115}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 20f710c5edfce534696d6814b28bb766, type: 3}
+  m_Script: {fileID: 11500000, guid: a7a8b90c36d79c246aab051870b3f8f5, type: 3}
   m_Name: 
   m_EditorClassIdentifier:
