+++ Assets/Prefabs/Managers/SpawnManager.prefab
@@ -16,7 +16,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &3807394749865851133
 Transform:
   m_ObjectHideFlags: 0
