--- Assets/Scripts/CurrentAttackableObjects.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dec7771bf9d127b49be23087f681e362
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
