--- Assets/Scripts/FarmHouseHealth.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8378f9831f786794bb13f39fc5943160
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
