+++ Assets/Scripts/GameManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 900d1e96de765434ba366e263a111eb5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
