+++ Assets/Scripts/GameManager/CurrentAttackableObjects.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2ff8cff738eb82245a5894bc9ca530e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
