+++ Assets/Scripts/GameManager/Pause.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a7a8b90c36d79c246aab051870b3f8f5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
