+++ Assets/Scripts/GameManager/SwitchBetweenAttackAndSpawnMode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f77ad97dc2a7d9741a95e7d66aad4fab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
