+++ Assets/Scripts/GameManager/SwitchBetweenRoundModeandUpgradeMode.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30bb6efb377e23849ada76f03f6d2adc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
