--- Assets/Scripts/Health.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9e4d879939308ba4db63f0f7234f13de
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
