+++ Assets/Scripts/Health.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3ef5df650a82e12459de4ef1bd3439f8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
