+++ Assets/Scripts/Health/FarmHouseHealth.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ea0f5985ed87300408b3b62c4ffaaf2d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
