+++ Assets/Scripts/Health/Health.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7bea43d7daa619649ba042ceb8d08bb9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
