+++ Assets/Scripts/UI/UpgradeMenu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2ea70aed30f32744db1bcc724a3eba32
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
