--- Assets/SwitchBetweenAttackAndSpawnMode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 20f710c5edfce534696d6814b28bb766
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
