--- Assets/SwitchBetweenRoundModeandUpgradeMode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4b1cb5385394e7548bb97671faff7023
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
